bug fixes from find bugs
git-svn-id: https://subversion.assembla.com/svn/smart-cs/trunk@8039 37136b1e-4cb1-4992-8394-573b48a9bdce

+++ source/extensions/i2/org.wcs.smart.i2/shared/org/wcs/smart/i2/model/IntelEntity.java
@@ -534,7 +534,7 @@ public class IntelEntity extends UuidItem implements IIntelAuditItem{
 					Double d = Double.parseDouble(ex[2]);
 					
 					if (op == Operator.EQUALS) {
-						match = value.getNumberValue() == d; 
+						match = value.getNumberValue().doubleValue() == d; 
 					}else if (op == Operator.LESSTHAN) {
 						match = value.getNumberValue() < d;
 					}else if (op == Operator.LESSTHANEQUALS) {
@@ -544,7 +544,7 @@ public class IntelEntity extends UuidItem implements IIntelAuditItem{
 					}else if (op == Operator.GREATERTHANEQUALS) {
 						match = value.getNumberValue() >= d;
 					}else if (op == Operator.NOTEQUALS) {
-						match = value.getNumberValue() != d;
+						match = value.getNumberValue().doubleValue() != d;
 					}else {
 						getDmAttributeListItem().setIsActive(defaultActive);
 						return;
